# docs/swagger.yaml
basePath: /api
definitions:
...
  recommend.ContextRequest:
    properties:
      latitude:
        type: number
      longitude:
        type: number
      timestamp:
        type: integer
    required:
    - latitude
    - longitude
    type: object
  recommend.ContextResponse:
    properties:
      location:
        $ref: '#/definitions/recommend.LocationInfo'
      time:
        $ref: '#/definitions/recommend.TimeInfo'
      weather:
        $ref: '#/definitions/recommend.WeatherInfoResponse'
    type: object
  recommend.LocationInfo:
    properties:
      city:
        type: string
      district:
        type: string
      province:
        type: string
    type: object
  recommend.RecommendRecipeItem:
    properties:
      category:
        type: string
      description:
        type: string
      difficulty:
        type: integer
      id:
        type: string
      image_path:
        type: string
      name:
        type: string
      reason:
        description: 该菜品的推荐理由
        type: string
      tags:
        $ref: '#/definitions/types.TagsGrouped'
      total_time_minutes:
        type: integer
    type: object
  recommend.RecommendRequest:
    properties:
      exclude_ids:
        description: 排除的菜谱 ID（换一批时传入已推荐的）
        items:
          type: string
        type: array
      latitude:
        type: number
      longitude:
        type: number
      timestamp:
        type: integer
    required:
    - latitude
    - longitude
    type: object
  recommend.RecommendResponse:
    properties:
      recipes:
        items:
          $ref: '#/definitions/recommend.RecommendRecipeItem'
        type: array
      remaining:
        description: 今日剩余推荐次数
        type: integer
      summary:
        description: LLM 生成的一句话整体评价
        type: string
    type: object
  recommend.TimeInfo:
    properties:
      day_of_week:
        type: integer
      hour:
        type: integer
      meal_time:
        type: string
      season:
        type: string
      timestamp:
        type: integer
    type: object
  recommend.WeatherInfoResponse:
    properties:
      humidity:
        type: integer
      icon:
        type: string
      temperature:
        type: number
      weather:
        type: string
    type: object
  tag.CreateTagRequest:
    properties:
      label:
        description: 显示名称
        type: string
      recipe_id:
        description: 菜谱ID（可选，为空时创建选项）
        type: string
      type:
        description: 标签类型
        type: string
      value:
        description: 标签值
        type: string
    required:
    - label
    - type
    - value
    type: object
  tag.TagValueResponse:
    properties:
      label:
        type: string
      value:
        type: string
    type: object
  tag.UpdateTagRequest:
    properties:
      label:
        type: string
    required:
    - label
    type: object
...
  types.TagsGrouped:
    properties:
      cuisines:
        items:
          type: string
        type: array
      flavors:
        items:
          type: string
        type: array
      scenes:
        items:
          type: string
        type: array
    type: object
...
host: localhost:18000
info:
  contact: {}
  description: 菜谱管理系统后端 API
  title: Choosy API
  version: "1.0"
paths:
...
  /api/recommend:
    post:
      consumes:
      - application/json
      description: 根据地理位置、天气、时间等因素智能推荐菜谱（基于 LLM，支持用户个性化）
      parameters:
      - description: 推荐请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/recommend.RecommendRequest'
      - default: 6
        description: 返回数量
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/recommend.RecommendResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 获取智能推荐菜谱
      tags:
      - recommend
  /api/recommend/context:
    post:
      consumes:
      - application/json
      description: 根据经纬度获取位置、天气、时间信息
      parameters:
      - description: 上下文请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/recommend.ContextRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/recommend.ContextResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 获取推荐上下文
      tags:
      - recommend
...
  /api/tags:
    get:
      description: 获取标签列表，支持通过查询参数过滤。所有类型（cuisine/flavor/scene/taboo/allergy）都统一使用此接口
      parameters:
      - description: '标签类型: cuisine/flavor/scene/taboo/allergy'
        in: query
        name: type
        type: string
      - description: 菜谱ID（获取特定菜谱的标签，仅用于 cuisine/flavor/scene）
        in: query
        name: recipe_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/tag.TagValueResponse'
            type: array
      summary: 获取标签列表
      tags:
      - tags
    post:
      consumes:
      - application/json
      description: 创建标签或选项。如果 recipe_id 为空，创建选项（仅支持 taboo/allergy）；如果 recipe_id 不为空，创建菜谱标签
      parameters:
      - description: 标签信息
        in: body
        name: tag
        required: true
        schema:
          $ref: '#/definitions/tag.CreateTagRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/tag.TagValueResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 创建标签
      tags:
      - tags
  /api/tags/{type}:
    get:
      description: 支持所有类型：cuisine/flavor/scene/taboo/allergy
      parameters:
      - description: '标签类型: cuisine/flavor/scene/taboo/allergy'
        in: path
        name: type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/tag.TagValueResponse'
            type: array
      summary: 根据类型获取标签
      tags:
      - tags
  /api/tags/{type}/{value}:
    delete:
      description: 删除标签或选项
      parameters:
      - description: 标签类型
        in: path
        name: type
        required: true
        type: string
      - description: 标签值
        in: path
        name: value
        required: true
        type: string
      - description: 菜谱ID（删除菜谱标签时必填，删除选项时留空）
        in: query
        name: recipe_id
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 删除标签
      tags:
      - tags
    put:
      consumes:
      - application/json
      description: 更新标签或选项的显示名称
      parameters:
      - description: 标签类型
        in: path
        name: type
        required: true
        type: string
      - description: 标签值
        in: path
        name: value
        required: true
        type: string
      - description: 菜谱ID（更新菜谱标签时必填，更新选项时留空）
        in: query
        name: recipe_id
        type: string
      - description: 更新内容
        in: body
        name: tag
        required: true
        schema:
          $ref: '#/definitions/tag.UpdateTagRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/tag.TagValueResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 更新标签
      tags:
      - tags
...
securityDefinitions:
  Bearer:
    description: 输入 "Bearer {token}"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"